Rename scheme to encoding
