+++ encoding/json/register_test.go
