+++ examples/json/server/main.go
@@ -24,7 +24,7 @@ import (
 	"fmt"
 
 	"github.com/yarpc/yarpc-go"
-	"github.com/yarpc/yarpc-go/scheme/json"
+	"github.com/yarpc/yarpc-go/encoding/json"
 	"github.com/yarpc/yarpc-go/transport"
 	"github.com/yarpc/yarpc-go/transport/http"
